+++ RELEASE_NOTES.md
@@ -1,5 +1,3 @@
 # Release Notes
 
-- remove santizer
-- remove musllinux temporalily to be built
-- update to adapt cookie like sk-ant-sidxx-
+- feat: replace RequireXApiKeyAuth with RequireFlexibleAuth fully to support bear token